backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
djoser
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
xhtml2pdf
channels
redis
django-redis
channels-redis
daphne
django-auth-adfs
pandas
openpyxl
django-model-utils
PyPDF2
whitenoise
django-simple-history
python-pptx
